strict: Fix bitrate fields in some clips
There was a bug in qtmux that wrote the bitrate and max. bitrate in the
wrong order. These files were generated before this bug was fixed. This
change just fixes the btrt atom in the MP4 header so that the bitrate
and max. bitrate (which is not set) are in the right order.